GenAi-week04_project01/requirements.txt
streamlit
langchain
langchain-huggingface
huggingface_hub
python-dotenv
pypdf